+++ admin/js.js
@@ -35,4 +35,28 @@ jQuery(document).ready(function($) {
 		button.next().val('');
 		button.hide().prev().html(downloadlistAdminJsVars.lbl_upload_button).addClass('button button-primary');
 	});
+
+	// save to hide transient-messages via ajax-request
+	$('div[data-dismissible] button.notice-dismiss').on('click',
+		function (event) {
+			event.preventDefault();
+			let $this = $(this);
+			let attr_value, option_name, dismissible_length, data;
+			attr_value = $this.closest('div[data-dismissible]').attr('data-dismissible').split('-');
+
+			// Remove the dismissible length from the attribute value and rejoin the array.
+			dismissible_length = attr_value.pop();
+			option_name = attr_value.join('-');
+			data = {
+				'action': 'downloadlist_dismiss_admin_notice',
+				'option_name': option_name,
+				'dismissible_length': dismissible_length,
+				'nonce': downloadlistAdminJsVars.dismiss_nonce
+			};
+
+			// run ajax request to save this setting
+			$.post(downloadlistAdminJsVars.ajax_url, data);
+			$this.closest('div[data-dismissible]').hide('slow');
+		}
+	);
 });