+++ templates/list-end.php
@@ -2,6 +2,7 @@
 /**
  * Template for ending list.
  *
+ * @version: 3.7.0
  * @package download-list-block-with-icons
  */